lab14/requirements.txt
alembic==1.8.1
bcrypt==4.0.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
phonenumbers==8.12.57
SQLAlchemy==1.4.42
Werkzeug==2.2.2
WTForms==3.0.1
pytest==7.2.0

Pillow~=9.3.0
PyJWT~=2.6.0